changed long to double in trip duration

+++ src/main/java/com/group_c/velo/service/TripService.java
@@ -6,5 +6,5 @@ public interface TripService {
     Trip getTrip(Long id);
     void deleteTrip(Long id);
     Trip saveTrip(Trip trip);
-    Long getDuration(Long id);
+    Double getDuration(Long id);
 }
